+++ changelog.md
@@ -38,5 +38,6 @@
 * Fixed lantern blocks not pushable by pistons
 * Fixed fireproof timber frames, wooden grates & lattices items.
 * Fixed [Trapdoor textures do not have rotation textures based on placement direction](https://github.com/Brandcraf06/Blockus/issues/332).
-
+* _Fix 2.9.4_ : Not compatible with 1.21
+* _Fix 2.9.4_ : Waxed Copper Bricks & Copper Tuff Bricks stonecutting recipes